Move Backup logic to separate fragment
