+++ travel-assistance/src/main/java/com/jraska/pwmd/travel/backup/BackupProgressDialog.java
@@ -58,6 +58,5 @@ public class BackupProgressDialog extends DialogFragment {
     return progressDialog;
   }
 
-
   //endregion
 }
